requirements.txt
flask==3.1.0
flask-cors==5.0.1
psycopg[binary]>=3.2.10
python-dotenv==1.1.0
